DP-142-Move the petition service to the logic folder
