+++ src/app/logic/committee/getPetitionsActiveService.service.ts
@@ -9,9 +9,10 @@ import {
   Subject,
   tap,
 } from 'rxjs';
-import { PetitionService } from 'src/app/features/new-petition/petition-service/petition.service';
+
 import { FilterData, Result } from 'src/app/shared/models/exports';
 import { ResponsePetition } from 'src/app/shared/models/petition/response-petition';
+import { PetitionService } from '../petition/exports';
 
 @Injectable({
   providedIn: 'root',
