+++ src/app/logic/petition/edit-petition-issue.service.ts
@@ -10,8 +10,8 @@ import {
   tap,
 } from 'rxjs';
 
-import { PetitionService } from 'src/app/features/new-petition/petition-service/petition.service';
 import { IssuePetitionData, Result } from 'src/app/shared/models/exports';
+import { PetitionService } from './exports';
 
 @Injectable({
   providedIn: 'root',
