+++ src/app/logic/petition/edit-petition.service.ts
@@ -9,9 +9,10 @@ import {
   Subject,
   tap,
 } from 'rxjs';
-import { PetitionService } from 'src/app/features/new-petition/petition-service/petition.service';
+
 import { Result } from 'src/app/shared/models/exports';
 import { ResponsePetition } from 'src/app/shared/models/petition/response-petition';
+import { PetitionService } from './exports';
 
 @Injectable({
   providedIn: 'root',
