+++ src/app/logic/petition/exports.ts
@@ -4,3 +4,4 @@ export * from './step-indicator.service';
 export * from './edit-petition-candidate.service';
 export * from './edit-petition-issue.service';
 export * from './edit-petition.service';
+export * from './petition.service';
